Client integration

+++ include/files.h
@@ -12,4 +12,4 @@ struct File_d** getDirectory(const char* path);
 
 long getFileSize(FILE* filePtr);
 
-FILE* getFilePtrFromDir(const struct File_d** dir, unsigned short num);
+FILE* getFilePtrFromDir(const struct File_d** dir, unsigned int num);
